Segregated volatile from fixed data
Allows the volatile data fields in the nodes/edges/vertexes structures
to be updated without updating gen_network.py as in C fields that are
missing in a statically initialization declaration are ignored.

Later this will be useful for separate data into rom and ram sections if
we run out of ram.